atoi

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -13,10 +13,10 @@ int _atoi(char *s)
 	{
 		if (s[i] == '-')
 			sign *= -1;
-		else if (s[i] >= 0 && s[i] <= 9)
+		else if ((int)s[i] >= 0 && (int)s[i] <= 9)
 		{
 			result *= 10;
-			result += (s[i] - 0);
+			result += ((int)s[i] - 0);
 		}
 		else if (result > 0)
 			break;
